# advanced/challenge-1/task1/server/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.21.2",
    "title": "REST API",
    "description": ""
  },
  "host": "",
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "",
      "description": ""
    }
  ],
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "AuthToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "The token for authentication"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth/signin": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/schedules/add": {
      "post": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/schedules/{id}": {
      "put": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/schedules/driver_schedules": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "type": "string"
          },
          {
            "name": "to",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/schedules/bus_schedules": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "type": "string"
          },
          {
            "name": "to",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/schedules/top_drivers_schedules": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/users/add": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "fullname": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/users/{id}": {
      "put": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User ID.",
            "value": 1
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "fullname": {
                  "example": "any"
                },
                "role_id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "security": [
    {
      "AuthToken": []
    }
  ]
}